requirements.txt
pygame
pyopengl
numpy
noise

